+++ sphinxcontrib/needs/functions/__init__.py
@@ -2,6 +2,6 @@
 from sphinxcontrib.needs.functions.functions import register_func, execute_func, resolve_dynamic_values, \
     find_and_replace_node_content, FunctionParsingException
 
-from sphinxcontrib.needs.functions.common import test, copy, check_linked_values
+from sphinxcontrib.needs.functions.common import test, copy, check_linked_values, calc_sum
 
-needs_common_functions = [test, copy, check_linked_values]
+needs_common_functions = [test, copy, check_linked_values, calc_sum]