+++ sphinxcontrib/needs/needs.py
@@ -310,7 +310,7 @@ def setup(app):
     # This should be called last, so that need-styles can override styles from used libraries
     app.connect('env-updated', install_styles_static_files)
 
-    return {'version': '0.3.0'}  # identifies the version of our extension
+    return {'version': '0.3.1'}  # identifies the version of our extension
 
 
 def visitor_dummy(*args, **kwargs):