+++ docs/changelog.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+0.3.2
+-----
+
+   * Bugfix: Links to parts of needs (:ref:`need_part`) are now stored and presented as *links incoming* of target link.
+
 0.3.1
 -----
  * Improvement: Added dynamic function :ref:`check_linked_values`.